fix: Correction du bug de chargement des pages admin - Conversion vers Server Components
- Transformation des pages admin/mailboxes et admin/followup-templates en Server Components
- Vérification d'authentification côté serveur au lieu de redirections client-side
- Suppression du setTimeout workaround dans followup-templates
- Création de Client Components séparés pour les parties interactives
- Correction de l'erreur ESLint avec useCallback dans followup-templates-page-client
- Les pages se chargent maintenant correctement sans nécessiter de rafraîchissement manuel
